+++ package/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.4.0] - 2018-09-09
+- Added support for local validation of packages with unpublished dependencies.
 - Add new public API to test all embedded packages.
 - Validate that package dependencies won't cause major conflicts
 - Validate that package has a minimum set of tests.